Added controller and deployment files to operator
The controller will automatically keep the alertmanager config up-to-date based on the presence or absence of secrets for Pager Duty and Dead Man's Snitch.

It's in a working state, but lacks the ability to reload alertmanager in order for the config changes to take effect. Also lacking unit tests and metrics. I've opened issues to address those items:

https://github.com/openshift/configure-alertmanager-operator/issues/1
https://github.com/openshift/configure-alertmanager-operator/issues/2