Modified the default route behavior to use DMS
The default route in Alertmanager is set to "null" at install time. This commit modifies the behavior of configure-alertmanager-operator to set the default route to "watchdog", so that our receiver will be used to send responses to DMS by default. When the DMS secret is deleted, the Alertmanager default "null" route is restored.

This fixes: https://github.com/openshift/configure-alertmanager-operator/issues/7