SREP-1186: Consolidate to 1 SelectorSyncSet
This operator doesn't require a `Namespace` to be created and so it's
okay to remove it. Thus, it also makes sense to consolidate the
`SelectorSyncSet` objects to contain (for this operator) `Subscription`
and `CatalogSource`.

While the other `SelectorSyncSet` operators may not be so lean as to
require just these two objects, they will also follow the same single
`SelectorSyncSet` pattern.

Signed-off-by: Lisa Seelye <lseelye@redhat.com>