Improve the operator image's score on quay.io
Current score for app-sre image (which was built 1 hour ago) is 16 High vulnerabilities.
https://quay.io/repository/app-sre/configure-alertmanager-operator?tab=tags

With this change the score changes to "Passed" (no vulnerabilities).
https://quay.io/repository/nmalik/configure-alertmanager-operator?tab=tags

To update the image I started with the Dockerfile from rbac-permissions-operator and incorporated what was needed and pushed to my personal quay.io account to see the score.  It still had a bad score.  I then updated the base image to ubi8.  I am not sure why rbac-permissions-operator is "Passed" with the ubi7 image.