Fix build and add podman support
On Fedora 31 I am unwilling to fight with docker, so I use podman.
Removed use of ENV for OPERATOR_PATH as it's not honored anymore for the second stage of the build.
Note tried using a combination of ARG and ENV in various ways and in all cases the path was not picked up, including with a rename of the ENV and/or ARG.