Convert from dep to to Go modules; and use Go 1.13
Motivation:

Newer versions of the operator-sdk CLI expect a Go modules project.
While the CLI isn't used directly here yet, we plan to add a CRD here
now and it will be more convenient to start using it for CRD generation.

Modification:

- `go mod init`

- Added replaces to go.mod as specified for the version of the
  operator-sdk specified in previous Gopkg.toml (v0.8.2).

- Add additional replace for prometheus/client_golang, as Go modules
  tries to use v1.0.0+ by default, which has a breaking change.

- `go mod init`

- `rm Gopkg.toml Gopkg.lock`

- `rm -r vendor/ && go mod vendor`

- Added `-mod=vendor` to GOFLAGS for Make so that it will use the
  vendor directory during builds, etc.

- Updated builder image in Dockerfile to use Go 1.13

Result:

- A lot of files are deleted from the vendor directory, since Go
modules only vendors packages that are used.
- GO111MODULE environment should be set to either on or auto, depending
on your configuration