Pass routes/receivers to createAlertManagerConfig
Before this, it would just create an empty slice of these; but now
we'll want it to take slices of them generated from the set of
AlertManagerConfiguration CRs.

I also updated the Scheme that's used by the fake client in the tests
in anticipation of needing to Get/List/etc instances of the new CRD
kind.