Remove TODO about ServiceKey field for PagerDuty
ServiceKey and RoutingKey are mutually exclusive depending on the
version of the PagerDuty Events API that's being used. Since
RoutingKey was already in use before this CRD, that means that it's v2
that's being used. Hopefully we don't have any need to use v1 for some
reason.

More info:
https://www.pagerduty.com/docs/guides/prometheus-integration-guide/

Also, dropping this field allows us to ignore having to validate that
only one of them was set.