Make golangci-lint pass
A prior commit bootstrapped this repository with the boilerplate
golang-osd-operator convention, which uses golangci-lint in place of
gofmt and go vet. This commit addresses the errors called out by
golangci-lint. Most of the fixes are legit, but a couple are spoofed,
with TODOs added, to be addressed later.

Jira: [OSD-4943](https://issues.redhat.com/browse/OSD-4943)