Boilerplate: Update to 263d5a8bd59a4bd1a0473b62d7e73a63c7002be0
Conventions:
- openshift/golang-osd-operator: Update
---
https://github.com/openshift/boilerplate/compare/d22b1886428528bac433e154f219b9c91437edd0...263d5a8bd59a4bd1a0473b62d7e73a63c7002be0

commit: 28da926566d421d5973b8579e7cf6b21b3df99d2
author: Naveen Malik
Don't break lint if there is no yaml to validate.

Make targets pass in files to validate but not all repos have yaml that needs validating.
In that case the yaml validation fails because it prints usage and exits with 1.

This PR simply changes the exit code to 0 so it's not going to fail.

There is no great alternative to this, we don't know where yaml might be or if there might be a directory to pass in etc.  Proposing we simply don't fail without arguments.

commit: 7c8928c70ec5dd216516cbe3d2c59a750f0eef57
author: Arjun Naik
Added a make target to build and push the boilerplate image

Signed-off-by: Arjun Naik <anaik@redhat.com>