Clean up app-interface targets
This PR rolls up a handful of related things.
- Boilerplate update commit 9c92f13 removed the `build` target. It was
just an alias for `docker-build`, so we fix the deploy script to use
that name instead.
- The [standardized app-interface configuration](https://gitlab.cee.redhat.com/service/app-interface/-/merge_requests/11165)
uses `make build-push` intsead of a hardcoded path to a script.
Boilerplate has that target hitting hack/app_sre_build_deploy.sh until
we can pull the logic into boilerplate itself. So for now, we rename
build/deploy.sh accordingly.
- PR checks are all being done in prow as of
https://github.com/openshift/release/pull/12412. We've [removed](https://gitlab.cee.redhat.com/service/app-interface/-/merge_requests/11164)
the ci-ext jenkins job that does the same thing via build/check_pr.sh,
so we can remove that script.