Fixes typo in alert silence causing error in cluster consoles
An extra ":" in one of the alert silences is causing the alertmanager
operator to fail to reconcile, with an "invalid label name" error, which
causes an alert/warning to be visible in the cluster console.

This removes the extra ":", which causes the label name to fail
validation via the Prometheus label name regex, from the string.

REF: https://issues.redhat.com/browse/OSD-8717
REF: https://github.com/openshift/configure-alertmanager-operator/commit/9bd5a951327d2c565d845acc352441ad601a9952

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>